// frontend/src/types/rate.ts
export interface Rate {
    _id: string;
    Description: string;
    Last: number;
    High: number;
    Low: number;
    Bid: number;
    Ask: number;
    DailyChange: number;
    DailyChangePercent: number;
    WeeklyChange: number;
    WeeklyChangePercent: number;
    MonthlyChange: number;
    MonthlyChangePercent: number;
    YearlyChange: number;
    YearlyChangePercent: number;
    Direction: number;
    DailyDirection: number;
    Volatility: number;
    PreviousClose: number;
    PreviousWeekClose: number;
    PreviousMonthClose: number;
    PreviousYearClose: number;
    WOWPreviuosClose: number;
    MOMPreviousClose: number;
    YOYPreviousClose: number;
    LegacyCode: string;
    Code: string;
    DateTime: bigint;
    WOWLow: number;
    WOWHigh: number;
    MOMLow: number;
    MOMHigh: number;
    YOYLow: number;
    YOYHigh: number;
    WTDLow: number;
    WTDHigh: number;
    WTDLowDate: bigint;
    WTFHighDate: bigint;
    YTDLow: number;
    YTDHigh: number;
    YTDLowDate: bigint;
    YTDHighDate: bigint;
    pageID: string;
}